refactor(api): improve readability of `requireDirectCount` fn
